Add resources to rotation MP (#86)
* Working Rotation MP with resources. Need to handle cons weekend and fix dynamic.

* Rolled back on TotalWeekend consumption. Too much overhead to handle long weekend in init.

* handle resource and dynamic constraints in rotation MP.

* Working management of the resources Shift duration and maxweekend in the rotation problem.

+++ src/solvers/StochasticSolver.h
@@ -246,9 +246,9 @@ class StochasticSolver : public Solver {
   // Evaluation
   vector2D<Solver *> pEvaluationSolvers_;
   std::vector<Solver *> pReusableEvaluationSolvers_;
-  std::vector<std::map<double, std::set<int> > >
+  std::vector<std::map<int, std::set<int> > >
       schedulesFromObjectiveByEvaluationDemand_;
-  std::vector<std::map<double, std::set<int> > >
+  std::vector<std::map<int, std::set<int> > >
       schedulesFromObjectiveByEvaluationDemandGreedy_;
   // Scores
   std::vector<double> theScores_;
